Update best_f1.txt(Had to change to default of 0)
Realized there is going to be an issue with our training loop checking this file to see if our Validation f1 score is greater than the score in the file.  Since we could not upload the best model checkpoint it will throw an error when trying to resume from the checkpointed model since it doesn't exist.